feat(sign_common.js): make sign_common.js file for common variables & func and modify checkAllInput on signup.js
