+++ src/pages/signin/index.html
@@ -21,7 +21,7 @@
       type="text/css"
       href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css"
     />
-    <script src="./signin.js" defer></script>
+    <script type="module" src="./signin.js" defer></script>
     <title>signin</title>
   </head>
   <body>
