+++ src/pages/signup/index.html
@@ -21,7 +21,7 @@
       type="text/css"
       href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css"
     />
-    <script src="./signup.js" defer></script>
+    <script type="module" src="./signup.js" defer></script>
     <title>signup</title>
   </head>
   <body>
