+++ src/pages/signup/signup.js
@@ -1,15 +1,23 @@
+import {
+  $email,
+  $password,
+  $form,
+  $pwInvisible,
+  $emailErrorMsg,
+  $passwordErrorMsg,
+  showEmailError,
+  deleteEmailError,
+  showPasswordError,
+  deletePasswordError,
+  togglePwVisibility,
+} from "/Weekly_mission/1-Weekly-Mission/src/assets/js/sign_common.js";
+
 // HTML tag를 담는 변수는 앞에 $를 붙임
-const $email = document.querySelector("#email");
-const $password = document.querySelector("#password");
 const $doubleCheckPw = document.querySelector("#doublecheck-pw");
-const $form = document.querySelector("form");
-const $pwInvisible = document.querySelector(".password-invisible");
 const $doubleCheckPwInvisible = document.querySelector(
   ".doublecheck-pw-invisible"
 );
 
-const $emailErrorMsg = document.createElement("div");
-const $passwordErrorMsg = document.createElement("div");
 const $doubleCheckPwErrorMsg = document.createElement("div");
 $emailErrorMsg.classList.add("email-error-msg");
 $passwordErrorMsg.classList.add("password-error-msg");
@@ -18,43 +26,8 @@ $email.after($emailErrorMsg);
 $password.after($passwordErrorMsg);
 $doubleCheckPw.after($doubleCheckPwErrorMsg);
 
-// togglePwVisibility() 에서 쓰임
-let togglePwVisible = false;
 // toggleDoubleCheckPwVisibility() 에서 쓰임
 let toggleDoubleCheckPwVisible = false;
-
-const showEmailError = (type) => {
-  if (type === "void") {
-    $emailErrorMsg.textContent = "이메일을 입력해주세요.";
-  } else if (type === "typo") {
-    $emailErrorMsg.textContent = "올바른 이메일 주소가 아닙니다.";
-  } else if (type === "already") {
-    $emailErrorMsg.textContent = "이미 사용 중인 이메일입니다.";
-  }
-  $email.classList.add("error-border");
-};
-
-const deleteEmailError = () => {
-  $emailErrorMsg.textContent = "";
-  $email.classList.remove("error-border");
-};
-
-const showPasswordError = (type) => {
-  if (type === "again") {
-    $passwordErrorMsg.textContent =
-      "비밀번호는 영문, 숫자 조합 8자 이상 입력해 주세요.";
-  } else if (type === "diff") {
-    $passwordErrorMsg.textContent = "비밀번호가 일치하지 않아요.";
-    $doubleCheckPwErrorMsg.textContent = "비밀번호가 일치하지 않아요.";
-    $doubleCheckPw.classList.add("error-border");
-  }
-  $password.classList.add("error-border");
-};
-
-const deletePasswordError = () => {
-  $passwordErrorMsg.textContent = "";
-  $password.classList.remove("error-border");
-};
 
 const deleteDoubleCheckPwError = () => {
   $doubleCheckPwErrorMsg.textContent = "";
@@ -127,32 +100,21 @@ const checkAllInput = () => {
   if (
     $emailErrorMsg.textContent === "" &&
     $passwordErrorMsg.textContent === "" &&
-    $doubleCheckPwErrorMsg.textContent === ""
+    $doubleCheckPwErrorMsg.textContent === "" &&
+    $email.value.length > 0 &&
+    $password.value.length > 0 &&
+    $doubleCheckPw.value.length > 0
   ) {
     // error msg 모두 처리 시 'folder/' 로 이동
     window.location.href =
-      "/Weekly_mission/1-Weekly-Mission/src/assets/pages/folder/index.html";
-  }
-};
-
-// 비밀번호란의 비번 보이기 토글버튼
-const togglePwVisibility = () => {
-  if (!togglePwVisible) {
-    // 비밀번호 보이게 하기
-    $pwInvisible.setAttribute(
-      "src",
-      "/Weekly_mission/1-Weekly-Mission/src/assets/images/svg/eye-on.svg"
-    );
-    $password.setAttribute("type", "text");
-    togglePwVisible = true;
-  } else {
-    // 비밀번호 가리기
-    $pwInvisible.setAttribute(
-      "src",
-      "/Weekly_mission/1-Weekly-Mission/src/assets/images/svg/eye-off.svg"
-    );
-    $password.setAttribute("type", "password");
-    togglePwVisible = false;
+      "/Weekly_mission/1-Weekly-Mission/src/pages/folder/index.html";
+  } else if (
+    $email.value === "" &&
+    $password.value === "" &&
+    $doubleCheckPw.value === ""
+  ) {
+    showEmailError("void");
+    showPasswordError("void");
   }
 };
